fix deploy problems
