+++ .eslintrc.js
@@ -1,8 +1,3 @@
 module.exports = {
-  extends: ["@mate-academy/eslint-config", "plugin:cypress/recommended"],
-  parser: "babel-eslint",
-  parserOptions: {
-    sourceType: "module",
-    allowImportExportEverywhere: true,
-  },
+  extends: ['@mate-academy/eslint-config', 'plugin:cypress/recommended'],
 };
