+++ src/index.html
@@ -5,12 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>2048</title>
     <link rel="stylesheet" href="./styles/main.scss">
-    <script
-      src="scripts/main.js"
-      type="module"
-      defer>
-    </script>
-
   </head>
   <body>
     <div class="container">
@@ -64,5 +58,6 @@
         <p class="message message-start">Press "Start" to begin game. Good luck!</p>
       </div>
     </div>
+    <script src="scripts/main.js" type="text/javascript"></script>
   </body>
 </html>
